api/requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
Werkzeug==3.1.3
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.10
python-dotenv==1.0.1
flask-cors==3.0.10